+++ apps/astarte_appengine_api/.formatter.exs
@@ -1,4 +1,4 @@
 [
-  import_deps: [:phoenix, :ecto, :skogsra],
+  import_deps: [:phoenix, :ecto, :skogsra, :stream_data],
   inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"]
 ]