+++ apps/astarte_appengine_api/config/test.exs
@@ -28,6 +28,9 @@ config :astarte_rpc, :amqp_connection, host: System.get_env("RABBITMQ_HOST") || 
 
 config :astarte_appengine_api, :rpc_client, MockRPCClient
 
+config :stream_data,
+  max_runs: 50
+
 config :astarte_appengine_api,
        :test_priv_key,
        {%{kty: :jose_jwk_kty_rsa},